Rubocop fix
