+++ app/decorators/instructor_decorator.rb
@@ -1,3 +1,6 @@
-class InstructorDecorator < BaseDecorator
+class InstructorDecorator < Draper::Decorator
+  include FullNameable
+  include AvatarResizable
+
   delegate_all
 end
