+++ app/decorators/student_decorator.rb
@@ -1,3 +1,6 @@
-class StudentDecorator < BaseDecorator
+class StudentDecorator < Draper::Decorator
   delegate_all
+
+  include FullNameable
+  include AvatarResizable
 end
