+++ app/javascript/stylesheets/navbar.scss
@@ -13,8 +13,7 @@
   padding-left: 10px;
 }
 
-.dropdown-menu .dropdown-item
-{
+.dropdown-menu .dropdown-item{
   height: 30px;
   &:hover{
     background: transparent;
