+++ app/views/instructors/profiles/show.html.erb
@@ -1,14 +1,14 @@
 <div class="main-body">
   <div class="row gutters-sm">
     <div class="col-md-4 mb-3">
-        <div class="card-body shadow p-3 mb-5 bg-body rounded">
-          <div class="d-flex flex-column align-items-center text-center ">
-            <%= image_tag(@student.img_resized, class: 'rounded-circle') %>
-            <div class="mt-3">
-              <h4><%= @student.full_name %></h4>
-            </div>
+      <div class="card-body shadow p-3 mb-5 bg-body rounded">
+        <div class="d-flex flex-column align-items-center text-center ">
+          <%= image_tag(@instructor.avatar_with_fallback, class: 'rounded-circle instr_img') %>
+          <div class="mt-3">
+            <h4><%= @instructor.full_name %></h4>
           </div>
         </div>
+      </div>
     </div>
     <div class="col-md-8">
       <div class="col-md-8">
@@ -18,7 +18,7 @@
               <h6 class="mb-0"> Name</h6>
             </div>
             <div class="col-sm-9 text-secondary">
-              <%= @student.name %>
+              <%= @instructor.name %>
             </div>
           </div>
           <hr>
@@ -27,7 +27,7 @@
               <h6 class="mb-0">Surname</h6>
             </div>
             <div class="col-sm-9 text-secondary">
-              <%= @student.surname %>
+              <%= @instructor.surname %>
             </div>
           </div>
           <hr>
@@ -36,7 +36,7 @@
               <h6 class="mb-0">Email</h6>
             </div>
             <div class="col-sm-9 text-secondary">
-              <%= @student.email %>
+              <%= @instructor.email %>
             </div>
           </div>
           <hr>
@@ -45,13 +45,13 @@
               <h6 class="mb-0">Your courses</h6>
             </div>
             <div class="col-sm-9 text-secondary">
-              <%= @enrollments.count %>
+              <%= @instructor.courses.count %>
             </div>
           </div>
           <hr>
           <div class="row">
             <div class="col-sm-12">
-              <%= link_to 'Edit', edit_student_path(@student), class: 'btn btn-secondary  btn-block rounded' %>
+              <%= link_to 'Edit', edit_instructors_profiles_path(@instructor), class: 'btn btn-secondary  btn-block rounded' %>
             </div>
           </div>
         </div>
