+++ app/views/shared/_navbar.html.erb
@@ -25,7 +25,7 @@
         <ul class="dropdown-menu">
           <li>
             <div class="dropdown-item">
-              <%= instructor_signed_in? ? link_to('Show details', instructors_path) : link_to('Show details', students_path) %>
+              <%= instructor_signed_in? ? link_to('Show details', instructors_profiles_path) : link_to('Show details', students_profiles_path) %>
             </div>
           </li>
           <li><hr class="dropdown-divider"></li>
